+++ client/src/pages/Assign Badge/AccountHolderForm.jsx
@@ -1,71 +1,15 @@
 /* eslint-disable max-len */
 /* eslint-disable no-unused-vars */
 import React, { useState } from "react";
-import { Form, Input, Radio } from "antd";
-import styled, { css } from "styled-components";
+import { Form } from "antd";
+import { useFormik, FormikProvider } from "formik";
+import { Form as FormikAntdForm, Input } from "formik-antd";
+import StyledForm from "./StyledAccountHolderForm";
 import Modal from "../../components/Modals/Modal";
 import useModal from "../../hooks/useModal";
 import PrimaryFormButton from "../../components/Button/PrimaryFormButton";
-import { theme } from "../../constants/theme";
 import SearchInput from "../../components/Input/SearchInput";
-const { display, body } = theme.typography.font.family;
-const { royalBlue } = theme.colors;
 const { TextArea } = Input;
-
-const StyledForm = styled(Form)`
-
-  width: 33%;
-  min-width: 250px;
-  margin: 2.5vh auto;
-
-  .ant-form-item-label label {
-    padding: 0;
-    width: 100%;
-    min-width: 250px;
-    font-weight: 800;
-    font-size: 16px;
-    font-family: ${body};
-  }
-
-.ant-input {
-  width: 100%;
-  min-width: 250px;
-  margin-bottom -7px;
-  padding: 4.5px;
-}
-
-.ant-form-item-control-input-content .jpEryk {
-    align-self: flex-end;
-    width: 148px;
-    font-size: 14px;
-}
-
-.modal {
-  border-radius: 10px;
-}
-h4 { padding: 0; margin: 0; }
-
-.ant-form-item-control-input-content {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  span {
-    text-align: left;
-  }
-}
-
-.ant-input-group-addon button {
-  border-top: hidden;
-  border-bottom: hidden;
-  border-right: hidden;
-  &:hover {
-    border: ${royalBlue}
-    border-top: solid;
-    border-bottom: solid;
-    border-right: solid;
-  }
-}
-`;
 
 export default function AccountHolderForm() {
   const { isShowing, toggle } = useModal();
@@ -100,36 +44,54 @@ export default function AccountHolderForm() {
     console.log("account holder submitting", accountValue, evidenceValue);
   };
 
+  const handleSubmit = values => {
+    alert(JSON.stringify(values, null, 2));
+  };
+
+  const formik = useFormik({
+    initialValues: {
+      accountHolder: "",
+      evidence: "",
+    },
+    onSubmit: values => {
+      handleSubmit(values);
+    },
+  });
+
   return (
-    <StyledForm
-      id = "create-badge-form"
-      {...formItemLayout}
-      layout={formLayout}
-      form={form}
-      initialValues={{
-        layout: formLayout,
-      }}
-      onValuesChange={onFormLayoutChange}
-    >
-      <Form.Item label="Select a FightPandemics account holder">
-        <SearchInput placeholder="Start typing names of the person you want to award badges to" forminputs={formInputs({ value })}/>
-      </Form.Item>
-      <Form.Item label="">
-        <span>Evidence</span>
-        <TextArea placeholder="Optionally provide evidence for the person who completed the task" forminputs={formInputs({ value })}/>
-      </Form.Item>
-      <Form.Item>
-        <PrimaryFormButton className="btn-right" type="primary" onClick={toggle} onSubmit={submitHandler()}>Assign Badge</PrimaryFormButton>
-        <Modal
-          isShowing={isShowing}
-          hide={toggle}
-          modalWidth={800}
-          title="Congratulations your badge has been assigned!"
-          buttonPrimary="Copy URL"
-          buttonSecondary="Link"
-          modalBodyText="Your badge has been sent to [recipient name]! They will receieve an email notification for their new badge. You now have xxx badges remaining."
-        />
-      </Form.Item>
-    </StyledForm>
+    <FormikProvider value={formik}>
+      <StyledForm
+        id = "create-badge-form"
+        {...formItemLayout}
+        layout={formLayout}
+        form={form}
+        initialValues={{
+          layout: formLayout,
+        }}
+        onValuesChange={onFormLayoutChange}
+      >
+        <FormikAntdForm>
+          <Form.Item label="Select a FightPandemics account holder">
+            <SearchInput name="accountHolder" placeholder="Start typing names of the person you want to award badges to" forminputs={formInputs({ value })}/>
+          </Form.Item>
+          <Form.Item label="">
+            <span>Evidence</span>
+            <TextArea name="evidence" placeholder="Optionally provide evidence for the person who completed the task" forminputs={formInputs({ value })}/>
+          </Form.Item>
+          <Form.Item>
+            <PrimaryFormButton className="btn-right" type="primary" onClick={toggle} onSubmit={handleSubmit}>Assign Badge</PrimaryFormButton>
+            <Modal
+              isShowing={isShowing}
+              hide={toggle}
+              modalWidth={800}
+              title="Congratulations your badge has been assigned!"
+              buttonPrimary="Copy URL"
+              buttonSecondary="Link"
+              modalBodyText="Your badge has been sent to [recipient name]! They will receieve an email notification for their new badge. You now have xxx badges remaining."
+            />
+          </Form.Item>
+        </FormikAntdForm>
+      </StyledForm>
+    </FormikProvider>
   );
 }
