+++ client/src/pages/Assign Badge/NonAccountHolderForm.jsx
@@ -1,88 +1,15 @@
 /* eslint-disable max-len */
 /* eslint-disable no-unused-vars */
 import React, { useState } from "react";
-import { Form, Input, Radio } from "antd";
-import styled from "styled-components";
+import { Form } from "antd";
+import { useFormik, FormikProvider } from "formik";
+import { Form as FormikAntdForm, Input, Radio } from "formik-antd";
+import StyledForm from "./StyledAccountHolderForm";
 import Modal from "../../components/Modals/Modal";
 import useModal from "../../hooks/useModal";
 import PrimaryFormButton from "../../components/Button/PrimaryFormButton";
-import { theme } from "../../constants/theme";
-import SearchInput from "../../components/Input/SearchInput";
-const { body } = theme.typography.font.family;
-const { royalBlue } = theme.colors;
-const { TextArea } = Input;
 
-const StyledForm = styled(Form)`
-
-  width: 33%;
-  min-width: 250px;
-  margin: 2.5vh auto;
-
-  button {
-    width: 100%;
-    margin: 0;
-  }
-  
-  span {
-    text-align: left;
-  }
-
-  .ant-form-item-label label {
-    padding: 0;
-    width: 100%;
-    min-width: 250px;
-    font-weight: 800;
-    font-size: 16px;
-    font-family: ${body};
-  }
-
-.ant-input {
-  width: 100%;
-  min-width: 250px;
-}
-.modal {
-  border-radius: 10px;
-}
-h4 { padding: 0; margin: 0; }
-
-.ant-form-item-control-input-content {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-}
-
-.ant-form-item-control-input-content .jpEryk {
-    align-self: flex-end;
-    width: 226px;
-}
-
-.radio {
-    display: flex
-    flex-direction: row
-    margin-bottom: 10px;
-    .ant-radio-wrapper {
-        justify-content: space-between;
-    }
-}
-
-.ant-radio-checked .ant-radio-inner{
-  border-color: ${royalBlue} !important ;
-}
-
-.ant-radio-checked:before{
-  background-color: ${royalBlue};
-}
-
-.ant-radio-checked .ant-radio-inner:after{
-  background-color: ${royalBlue};
-}
-
-.ant-radio:hover .ant-radio-inner {
-  border-color: ${royalBlue} ;
-}
-`;
-
-export default function AccountHolderForm() {
+export default function NonAccountHolderForm() {
   const { isShowing, toggle } = useModal();
   const [form] = Form.useForm();
   const [formLayout, setFormLayout] = useState("vertical");
@@ -122,49 +49,67 @@ export default function AccountHolderForm() {
     );
   };
 
+  const handleSubmit = values => {
+    alert(JSON.stringify(values, null, 2));
+  };
+
+  const formik = useFormik({
+    initialValues: {
+      name: "",
+      email: "",
+    },
+    onSubmit: values => {
+      handleSubmit(values);
+    },
+  });
+
   return (
-    <StyledForm
-      id = "create-badge-form"
-      {...formItemLayout}
-      layout={formLayout}
-      form={form}
-      initialValues={{
-        layout: formLayout,
-      }}
-      onValuesChange={onFormLayoutChange}
-    >
-      <Form.Item label= "Share link with a non-account holder">
-        <span>Name</span>
-        <Input placeholder="First and Last Name" forminputs={formInputs(value)} />
-      </Form.Item>
-      <Form.Item >
-        <Radio.Group className="radio" onChange={onChange} value={value} forminputs={formInputs({ value })}>
-          <Radio value={1}>I know the recipient&apos;s email </Radio>
-          <Radio value={2}>I don&apos;t know the recipient&apos;s email </Radio>
-        </Radio.Group>
-      </Form.Item>
+    <FormikProvider value={formik}>
+      <StyledForm
+        id = "create-badge-form"
+        {...formItemLayout}
+        layout={formLayout}
+        form={form}
+        initialValues={{
+          layout: formLayout,
+        }}
+        onValuesChange={onFormLayoutChange}
+      >
+        <FormikAntdForm>
+          <Form.Item label= "Share link with a non-account holder">
+            <span>Name</span>
+            <Input name="name" placeholder="First and Last Name" forminputs={formInputs(value)} />
+          </Form.Item>
+          <Form.Item >
+            <Radio.Group className="radio" onChange={onChange} value={value} forminputs={formInputs({ value })}>
+              <Radio value={1}>I know the recipient&apos;s email </Radio>
+              <Radio value={2}>I don&apos;t know the recipient&apos;s email </Radio>
+            </Radio.Group>
+          </Form.Item>
 
-      {value === 1
-        ? (
-          <Form.Item label="">
-            <span>Email</span>
-            <Input placeholder="Email of non-account holder recipient" />
-          </Form.Item>)
-        : null
-      }
+          {value === 1
+            ? (
+              <Form.Item label="">
+                <span>Email</span>
+                <Input name="email" placeholder="Email of non-account holder recipient" />
+              </Form.Item>)
+            : null
+          }
 
-      <Form.Item>
-        <PrimaryFormButton className="btn-right" type="primary" onClick={toggle} onSubmit={linkSubmitHandler()}>Generate Link to Share</PrimaryFormButton>
-        <Modal
-          isShowing={isShowing}
-          hide={toggle}
-          modalWidth={800}
-          title="Share Link with a non-account holder"
-          buttonPrimary="Copy URL"
-          buttonSecondary="Link"
-          modalBodyText="The badge link has been successfully generated for [name]. Please copy this link to send to them!"
-        />
-      </Form.Item>
-    </StyledForm>
+          <Form.Item>
+            <PrimaryFormButton className="btn-right" type="primary" onClick={toggle} onSubmit={handleSubmit}>Generate Link to Share</PrimaryFormButton>
+            <Modal
+              isShowing={isShowing}
+              hide={toggle}
+              modalWidth={800}
+              title="Share Link with a non-account holder"
+              buttonPrimary="Copy URL"
+              buttonSecondary="Link"
+              modalBodyText="The badge link has been successfully generated for [name]. Please copy this link to send to them!"
+            />
+          </Form.Item>
+        </FormikAntdForm>
+      </StyledForm>
+    </FormikProvider>
   );
 }
